requirements.txt
python-telegram-bot==20.3.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.87.0
google-auth-httplib2==0.1.0
pygsheets==2.0.6
